Modularize ranking showcase
